+++ packages/api-headless-cms/src/types.ts
@@ -630,7 +630,7 @@ export interface CmsSettings {
     /**
      * Last content model change. Used to cache GraphQL schema.
      */
-    contentModelLastChange: Date;
+    contentModelLastChange: string;
 }
 
 /**
@@ -648,10 +648,6 @@ export interface CmsSettingsContext {
          */
         get: () => Promise<CmsSettings | null>;
     };
-    /**
-     * Last content model change. Used to cache GraphQL schema.
-     */
-    contentModelLastChange: Date;
     /**
      * Gets settings model from the database.
      */
@@ -1808,8 +1804,9 @@ export interface CmsModelFieldToElasticsearchPlugin extends Plugin {
  * @category Storage
  */
 export interface CmsModelFieldToStoragePluginToStorageArgs {
-    field: CmsContentModelField;
     model: CmsContentModel;
+    entry: CmsContentEntry;
+    field: CmsContentModelField;
     context: CmsContext;
     /**
      * The value to be transformed to storage type from the original one.
@@ -1827,8 +1824,9 @@ export interface CmsModelFieldToStoragePluginToStorageArgs {
  * @category Storage
  */
 export interface CmsModelFieldToStoragePluginFromStorageArgs {
-    field: CmsContentModelField;
     model: CmsContentModel;
+    entry: CmsContentEntry;
+    field: CmsContentModelField;
     context: CmsContext;
     /**
      * The value to be transformed from storage type into the original one.