+++ packages/api-page-builder/src/graphql/crud/utils/defaults.ts
@@ -1,4 +1,4 @@
-import { PbContext } from "../../../types";
+import { PbContext } from "../../types";
 
 export default {
     db: {
@@ -25,12 +25,15 @@ export default {
     },
     es(context: PbContext) {
         const tenant = context.security.getTenant();
-        if (tenant) {
-            return {
-                index: tenant.id + "-page-builder"
-            };
+        if (!tenant) {
+            throw new Error("Tenant missing.");
         }
 
-        throw new Error("Tenant missing.");
+        const index = tenant.id + "-page-builder";
+        const prefix = process.env.ELASTIC_SEARCH_INDEX_PREFIX;
+        if (prefix) {
+            return { index: prefix + index };
+        }
+        return { index };
     }
 };