+++ packages/app-page-builder/src/editor/plugins/elements/social/twitter/index.tsx
@@ -1,11 +1,9 @@
 import React from "react";
 import styled from "@emotion/styled";
+import kebabCase from "lodash/kebabCase";
 import { Typography } from "@webiny/ui/Typography";
 import { validation } from "@webiny/validation";
-import {
-    createEmbedPlugin,
-    createEmbedSettingsPlugin
-} from "./../../utils/oembed/createEmbedPlugin";
+import { createEmbedPlugin, createEmbedSettingsPlugin } from "../../utils/oembed";
 import placeholder from "./placeholder.jpg";
 import { ReactComponent as LogoIcon } from "./twitter-brands.svg";
 import Accordion from "../../../elementSettings/components/Accordion";
@@ -14,6 +12,7 @@ import {
     ButtonContainer,
     SimpleButton
 } from "../../../elementSettings/components/StyledComponents";
+import { PbEditorElementPluginArgs } from "../../../../../types";
 
 declare global {
     interface Window {
@@ -30,51 +29,62 @@ const PreviewBox = styled("div")({
     }
 });
 
-export default () => [
-    createEmbedPlugin({
-        type: "twitter",
-        toolbar: {
-            title: "Tweet",
-            group: "pb-editor-element-group-social",
-            preview() {
-                return (
-                    <PreviewBox>
-                        <LogoIcon />
-                    </PreviewBox>
-                );
-            }
-        },
-        oembed: {
-            global: "twttr",
-            sdk: "https://platform.twitter.com/widgets.js",
-            init({ node }) {
-                window.twttr.widgets.load(node);
-            }
-        },
-        renderElementPreview({ width, height }) {
-            return <img style={{ width, height }} src={placeholder} alt={"Tweet"} />;
-        }
-    }),
-    createEmbedSettingsPlugin({
-        type: "twitter",
-        render({ Bind, submit }) {
+export default (args: PbEditorElementPluginArgs = {}) => {
+    const elementType = kebabCase(args.elementType || "twitter");
+    const defaultToolbar = {
+        title: "Tweet",
+        group: "pb-editor-element-group-social",
+        preview() {
             return (
-                <Accordion title={"Twitter"} defaultValue={true}>
-                    <>
-                        <Bind name={"source.url"} validators={validation.create("required,url")}>
-                            <InputField
-                                placeholder={"https://twitter.com/"}
-                                description={"Enter a Tweet URL"}
-                            />
-                        </Bind>
-                        <ButtonContainer>
-                            <SimpleButton onClick={submit}>
-                                <Typography use={"caption"}>Save</Typography>
-                            </SimpleButton>
-                        </ButtonContainer>
-                    </>
-                </Accordion>
+                <PreviewBox>
+                    <LogoIcon />
+                </PreviewBox>
             );
         }
-    })
-];
+    };
+
+    return [
+        createEmbedPlugin({
+            type: elementType,
+            toolbar:
+                typeof args.toolbar === "function" ? args.toolbar(defaultToolbar) : defaultToolbar,
+            create: args.create,
+            settings: args.settings,
+            oembed: {
+                global: "twttr",
+                sdk: "https://platform.twitter.com/widgets.js",
+                init({ node }) {
+                    window.twttr.widgets.load(node);
+                }
+            },
+            renderElementPreview({ width, height }) {
+                return <img style={{ width, height }} src={placeholder} alt={"Tweet"} />;
+            }
+        }),
+        createEmbedSettingsPlugin({
+            type: elementType,
+            render({ Bind, submit }) {
+                return (
+                    <Accordion title={"Twitter"} defaultValue={true}>
+                        <>
+                            <Bind
+                                name={"source.url"}
+                                validators={validation.create("required,url")}
+                            >
+                                <InputField
+                                    placeholder={"https://twitter.com/"}
+                                    description={"Enter a Tweet URL"}
+                                />
+                            </Bind>
+                            <ButtonContainer>
+                                <SimpleButton onClick={submit}>
+                                    <Typography use={"caption"}>Save</Typography>
+                                </SimpleButton>
+                            </ButtonContainer>
+                        </>
+                    </Accordion>
+                );
+            }
+        })
+    ];
+};