+++ packages/cli-plugin-scaffold-admin-app-module/template/src/views/graphql.ts
@@ -22,6 +22,11 @@ export const LIST_TARGETS = gql`
                         type
                     }
                 }
+                error {
+                    message
+                    code
+                    data
+                }
             }
         }
     }