+++ packages/cli-plugin-scaffold-graphql-service/template/src/resolvers/createTarget.ts
@@ -8,7 +8,7 @@ const createTarget = async (
     args: CreateTargetArgs,
     context: ApplicationContext
 ): Promise<ResolverResponse<Target>> => {
-    const { db, security } = context;
+    const { db, security, elasticSearch } = context;
     const { data } = args;
 
     /**
@@ -31,10 +31,27 @@ const createTarget = async (
         isNice: data.isNice === undefined ? false : data.isNice
     };
     /**
-     * Create the index name that is going to be used when streaming from DDB to Elasticsearch.
+     * Create, and check for existence, index name that is going to be used when streaming from DDB to Elasticsearch.
      * Can be removed if Elasticsearch is not used.
      */
-    const { index: esIndex } = utils.es(context);
+    const esConfig = utils.es(context);
+    try {
+        const { body: hasIndice } = await elasticSearch.indices.exists(esConfig);
+        if (!hasIndice) {
+            return new ErrorResponse({
+                message: "You must run the install mutation to create the Elasticsearch index.",
+                code: "ELASTICSEARCH_ERROR",
+                data: esConfig
+            });
+        }
+    } catch (ex) {
+        return new ErrorResponse({
+            message: ex.message || "Error while checking for Elasticsearch index existence.",
+            code: ex.code || "ELASTICSEARCH_ERROR",
+            data: ex.data
+        });
+    }
+    const { index: esIndex } = esConfig;
 
     /**
      * Primary key is always constructed out of the target.id and a fixed Target configuration.