+++ packages/cwp-template-aws/template/api/pulumi/dev/elasticSearch.ts
@@ -27,21 +27,51 @@ class ElasticSearch {
             }
         });
 
+        /**
+         * Domain policy defines who can access your Elasticsearch Domain.
+         * For details on Elasticsearch security, read the official documentation:
+         * https://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/security.html
+         */
         new aws.elasticsearch.DomainPolicy(`${domainName}-policy`, {
             domainName: this.domain.domainName.apply(v => `${v}`),
-            accessPolicies: {
-                Version: "2012-10-17",
-                Statement: [
-                    {
-                        Action: ["es:*"],
-                        Principal: {
-                            AWS: ["*"]
-                        },
-                        Effect: "Allow",
-                        Resource: this.domain.arn.apply(v => `${v}/*`)
-                    }
-                ]
-            }
+            accessPolicies: Promise.all([aws.getCallerIdentity({})]).then(
+                ([currentCallerIdentity]) => ({
+                    Version: "2012-10-17",
+                    Statement: [
+                        /**
+                         * Allow requests signed with current account
+                         */
+                        {
+                            Effect: "Allow",
+                            Principal: {
+                                AWS: currentCallerIdentity.accountId
+                            },
+                            Action: "es:*",
+                            Resource: this.domain.arn.apply(v => `${v}/*`)
+                        }
+                        /**
+                         * Uncomment the following `Allow` policy to allow access from specific IP address.
+                         * This will be useful for development purposes, when you want to access Kibana to inspect your data.
+                         *
+                         * If you need to setup proper user accounts for access to Kibana, you'll need to connect it to
+                         * Cognito User Pool. For instructions, see the official documentation:
+                         * https://docs.aws.amazon.com/elasticsearch-service/latest/developerguide/es-cognito-auth.html
+                         */
+
+                        // {
+                        //     Effect: "Allow",
+                        //     Principal: "*",
+                        //     Action: "es:*",
+                        //     Resource: this.domain.arn.apply(v => `${v}/!*`),
+                        //     Condition: {
+                        //         IpAddress: {
+                        //             "aws:SourceIp": "213.149.51.28/32"
+                        //         }
+                        //     }
+                        // }
+                    ]
+                })
+            )
         });
 
         /**