+++ packages/cli-plugin-scaffold-graphql-service/CHANGELOG.md
@@ -3,6 +3,61 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.2.0](https://github.com/webiny/webiny-js/compare/v5.2.0-beta.0...v5.2.0) (2021-03-19)
+
+**Note:** Version bump only for package @webiny/cli-plugin-scaffold-graphql-service
+
+
+
+
+
+# [5.2.0-beta.0](https://github.com/webiny/webiny-js/compare/v5.1.0...v5.2.0-beta.0) (2021-03-19)
+
+**Note:** Version bump only for package @webiny/cli-plugin-scaffold-graphql-service
+
+
+
+
+
+# [5.1.0](https://github.com/webiny/webiny-js/compare/v5.1.0-beta.1...v5.1.0) (2021-03-16)
+
+**Note:** Version bump only for package @webiny/cli-plugin-scaffold-graphql-service
+
+
+
+
+
+# [5.1.0-beta.1](https://github.com/webiny/webiny-js/compare/v5.1.0-beta.0...v5.1.0-beta.1) (2021-03-16)
+
+
+### Bug Fixes
+
+* **cli-plugin-scaffold-graphql-service:** es error message explained ([#1510](https://github.com/webiny/webiny-js/issues/1510)) ([106e3ac](https://github.com/webiny/webiny-js/commit/106e3acfd16255f8038d68c77e94d5e0c598460b))
+
+
+
+
+
+# [5.1.0-beta.0](https://github.com/webiny/webiny-js/compare/v5.0.0...v5.1.0-beta.0) (2021-03-15)
+
+
+### Bug Fixes
+
+* **cli-plugin-scaffold-graphql-service:** es index check on create ([#1509](https://github.com/webiny/webiny-js/issues/1509)) ([35e25de](https://github.com/webiny/webiny-js/commit/35e25de549aa4442f4e1c1cc4982dd9799a81af7))
+* update babel configs in templates ([b33096f](https://github.com/webiny/webiny-js/commit/b33096ff34ac2a745ead452f127a039a12c72aac))
+* **cli-plugin-scaffold-admin-app-module:** accessing non-objects error ([#1505](https://github.com/webiny/webiny-js/issues/1505)) ([1e2e36b](https://github.com/webiny/webiny-js/commit/1e2e36b907244ef8f844ca25f4db9228c300b870))
+* **cwp-template-aws:** add missing devDependencies ([#1496](https://github.com/webiny/webiny-js/issues/1496)) ([e340ff6](https://github.com/webiny/webiny-js/commit/e340ff6bb3cbda3159ddc17d1fe0631dde54084a))
+
+
+### Features
+
+* **app-headless-cms:** add support for content model form layouts ([#1507](https://github.com/webiny/webiny-js/issues/1507)) ([f2b65b0](https://github.com/webiny/webiny-js/commit/f2b65b0f593e83948b85832d3234e4d32793a868))
+* add the ability to use ~ for import paths within packages ([#1508](https://github.com/webiny/webiny-js/issues/1508)) ([0a17a6f](https://github.com/webiny/webiny-js/commit/0a17a6f7463a93ad7c0f5b3a7cd1d7dbfc5e0012))
+
+
+
+
+
 # [5.0.0](https://github.com/webiny/webiny-js/compare/v5.0.0-beta.5...v5.0.0) (2021-03-09)
 
 **Note:** Version bump only for package @webiny/cli-plugin-scaffold-graphql-service